More screen resize and center changes
Height and width of video will now properly be set using the video height scaled for the correct aspect ratio. Centering the window will take into account system UI elements like taskbars on the bottom and left side only (will be uniquely incorrect according to the operating system if not).