Build Improvements (#13)
* Change build configuration names

The build configurations had nothing to do with architecture so remove that from the name. Also add an architecture identifier to the updater.

* Fix licenses and readme not copying to output

* Create windows setup with innosetup

* Improve osx build

Fixes error on dotnet publish with osx and expands upon the .plist

* Remove updater from osx build

The updater does not work on osx with their special permissions. So checking for updates will open a web link to the latest release if an update is detected and yes is selected.

* Remove CopyUpdater path parameter